rename main level fakeHaplotreeCode.py to examples dir 	modified:   examples/fakeHaplotreeCode.py
